src/adaptive/requirements.txt
fastapi
uvicorn
groq
python-dotenv
mysql-connector-python
httpx